Move error handler and logger to subdirectory.
