+++ packages/coinstac-simulator/src/utils/handle-errors.js
