+++ packages/coinstac-simulator/src/utils/logger.js
